benchmark/requirements.txt
boto3>=1.34.0
psycopg2-binary>=2.9.9
kubernetes>=29.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
